feat: 활동 기수 api 컨트롤러 작성
