+++ src/main/java/com/tavemakers/surf/domain/activity/dto/acitveGeneration/request/ActiveGenerationUpdateReqDTO.java
@@ -3,7 +3,7 @@ package com.tavemakers.surf.domain.activity.dto.acitveGeneration.request;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
-public record ActiveGenerationPatchReqDTO(
+public record ActiveGenerationUpdateReqDTO(
         @NotNull
         @Min(1)
         Integer activeGeneration
